Update setup script to check and handle case when regular file exists

+++ env-setup.sh
@@ -18,21 +18,59 @@ do
         # Checks if target file in home directory is a symlink
         if [ -L "$home_dir/$target" ]
         then
-            printf "$target is already present in your home directory\n"
+            printf "$target is already a symlink in your home directory\n"
         else
-            printf "\nLinking $target\n"
-
-            # Links file/directory within dotfiles to home directory
-            ln -s $proj_dir/$target $home_dir/$target
-
-            # Success checks for previous command
-            if [ $? -eq 0 ]
+            if [ -f "$home_dir/$target" ]
             then
-                printf "Link successful\n"
-            else
-                printf "Linking was unsuccessful\n"
+                answer="Y"
+
+                printf "$target already exists in your home directory\n"
+                printf "Do you want to delete the file and create a symlink instead? The deletion is permanent! [Y/n] "
+                # Read user input
+                read reply
+
+                # If input is empty or null, replace with $answer, else reply remains the same
+                reply=${reply:=${answer,,}}
+
+                if [ ${reply,,} = "y" ] 
+                then
+                    # Deletes file in home directory
+                    rm -i $home_dir/$target
+                    
+                    # Checks if command is ran successfully and the target file does not exist
+                    if [ $? -eq 0 ] && [ ! -f "$home_dir/$target" ]
+                    then
+                        printf "\nDelete successful\n"
+
+                        printf "\nLinking $target\n"
+
+                        # Links file/directory within dotfiles to home directory
+                        ln -s $proj_dir/$target $home_dir/$target
+
+                        # Success checks for previous command
+                        if [ $? -eq 0 ]
+                        then
+                            printf "Link successful\n"
+
+                        else
+                            printf "Linking was unsuccessful\n"
+
+                        fi
+                    else
+                        printf "Delete failed!\n"
+
+                    fi
+
+                elif [ ${reply,,} = "n" ]
+                then
+                    printf "No changes have been made to $target\n"
+
+                else
+                    printf "\nInvalid input detected\n"
+                fi
             fi
         fi
     fi
 done
+printf "Exiting script...\n"
